Full-Stack Application (React, Docker, Python and more)/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
